refactor(liveobjects): extract ObjectUpdate model out of BaseRealtimeLiveObject
Move the ObjectUpdate sealed hierarchy (NoOp/MapUpdate/CounterUpdate) and its
noOp extension into a dedicated value/ObjectUpdate.kt, leaving BaseRealtimeLiveObject
focused on the base class. ObjectType stays with the base class as the broad object
discriminator used across the package.

Relocate the LiveMap-specific MapChange enum from the generic value package into
value.livemap alongside the other map types; it is a per-key diff type owned by
LiveMap, and value already depends on value.livemap (BaseRealtimeLiveObject uses
InternalLiveMap), so this introduces no new package coupling.

The sealed subclasses remain in the value package, so exhaustive when-handling
(e.g. tombstone) still compiles. Pure code move, no behaviour change.

Also drop a stale @Suppress(RedundantNullableReturnType) on
DefaultLiveMapInstance.get, whose nullable return is genuine.
